+++ src/components/main/MainContent.module.scss
@@ -1,0 +1,61 @@
+@use '../../styles/variable.module' as var;
+
+.postTweetContainer {
+  display: flex;
+  flex-direction: column;
+  height: calc(100% - var.$sidebar-height);
+}
+
+.avatar{
+  width: 50px;
+  height: 50px;
+  border-radius: 50%;
+}
+
+.postTweetForm {
+  height: 100%;
+
+}
+
+.postTweetContent {
+  height: 100%;
+  flex: 1;
+  padding: 16px 55px 16px 24px;
+}
+
+.contentGroup {
+  display: flex;
+  gap: 8px;
+  height: 100%;
+
+  .postTweetInputGroup {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    height: 100%;
+    width: 100%;
+  }
+  
+
+  .postTweetInput {
+    resize: none;
+    border: none;
+    box-shadow: none;
+    outline: none;
+    color: #6c757d;
+    font-size: 18px;
+    font-weight: 700;
+    width: 100%;
+    height: 100%;
+    overflow: scroll;
+    margin-top: 5px;
+  }
+  .footerArea {
+    display: flex;
+    justify-content: flex-end;
+  }
+  .footerButton {
+   width: 64px;
+  }
+  
+}
