+++ src/components/main/MainTweets.jsx
@@ -5,6 +5,8 @@ import AdminTweetsItem from 'components/admin/AdminTweetsItem'
 import Sidebar from 'UIcomponents/layouts/Sidebar'
 import { Outlet } from 'react-router-dom'
 import Layout from 'UIcomponents/layouts/Layout'
+import MainContent from './MainContent'
+
 const dummyData = [
   {
     id: '1',
@@ -116,12 +118,14 @@ export default function MainTweets () {
     <div className={`${style.tweetsContainer}`}>
       <header className={`${style.tweetsHeader}`}>
         <nav className={`${style.tweetsNavbar}`}>
-          <div className={`${style.navAvatarItem}`}><img className={`${style.navAvatar}`} src='https://picsum.photos/300/300?text=1' alt="" /></div>
           <div className={`${style.navTitle}`}>
             <h1 className={`${style.titleText}`}>首頁</h1>
           </div>
         </nav>
       </header>
+      <div className={`${style.tweetPostArea}`}>
+        <MainContent/>
+      </div>
       <main className={`${style.mainTweets}`}>
         {dummyData.map((data) => (
           <MainTweetsContent key={data.id} {...data}/>
