--- src/components/main/PostTweet.module.scss
@@ -1,9 +1,0 @@
-.avatar{
-  width: 50px;
-  height: 50px;
-  border-radius: 50%;
-}
-
-.contentGroup {
-  display: flex;
-}
