+++ src/styles/_variable.module.scss
@@ -1,0 +1,1 @@
+$sidebar-height: 52px;
